requirements.txt
yfinance
nsepython
groq
gTTS
matplotlib
plotly
moviepy
numpy
pandas
python-dotenv
pillow
requests
kaleido
mplfinance
bar_chart_race
pyttsx3
nselib
google-generativeai
pywin32
pip
